feat(day09): solution for part 2

+++ src/solutions/day09.h
@@ -3,4 +3,4 @@
 #include <vector>
 
 int64_t day09_part1(const std::vector<std::string> &input);
-int32_t day09_part2(const std::vector<std::string> &input);
+int64_t day09_part2(const std::vector<std::string> &input);
